refactor: move text related elements under `text` module
