+++ src/webdynpro/element/text/label.rs
@@ -1,6 +1,7 @@
 use std::{borrow::Cow, cell::OnceCell};
 
-use super::define_element_interactable;
+use crate::webdynpro::element::define_element_interactable;
+
 
 // TODO: Implement additional events and data
 define_element_interactable! {
